Fixed bug in buzzvm_function_call().
The bug was subtle. The function expects the stack to have the arguments pushed
in order. Internally, it looks for the closure corresponding to the passed
string, which leaves the closure on top of the stack. However, a call to
buzzvm_closure_call is done that requires the closure to be after the arguments,
not on top of them, in the stack. The happened when the closure was inserted
before the arguments. Rather than passing a pointer to pointer to buzzobj_t, the
function passed a straight pointer to buzzobj_t. This made it so memcpy would
copy the wrong data into the array slot - rather than the pointer to buzzobj_t
that stores the closure, it was copying the first 8 bytes of the closure data itself.